requirements.txt
Flask==2.3.3
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
